+++ lib/modules/test/test_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: test
-// Updated at: 2026-06-29 16:26:50
-// Commit: feat(test): add async test stream provider
+// Updated at: 2026-06-30 10:34:07
+// Commit: feat(test): integrate custom telemetry logger in test
 
 class testComponentState {
-  final String commitRef = '2407';
-  final String lastMessage = 'feat(test): add async test stream provider';
-  final DateTime timestamp = DateTime.parse('2026-06-29T16:26:50+05:00');
+  final String commitRef = '2421';
+  final String lastMessage = 'feat(test): integrate custom telemetry logger in test';
+  final DateTime timestamp = DateTime.parse('2026-06-30T10:34:07+05:00');
 }